requirements.txt
arrow==1.2.3
binaryornot==0.4.4
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cookiecutter==2.1.1
cycler==0.11.0
duckdb-engine
fonttools==4.39.2
idna==3.4
graphviz
imgkit
imbalanced-learn==0.10.1
importlib-resources==5.12.0
Jinja2==3.1.2
jinja2-time==0.2.0
jupysql
kaleido
keras==2.12.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
nbformat==5.8.0
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.14.1
ploomber
psycopg2-binary==2.9.5
pyarrow==11.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit_learn==1.2.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
statsmodels==0.13.5
sqlalchemy==2.0.7
sqlparse==0.4.3
sqlglot>=11.3.7
six==1.16.0
text-unidecode==1.3
urllib3==1.26.15
wincertstore==0.2
xgboost==1.7.4
zipp==3.15.0
pytest